data_preparation/requirements.txt
tensorflow
tensorflow_datasets==4.6.0
datasets 
huggingface_hub
numpy
Pillow
opencv-python
protobuf==3.20.*